test: first real test also run in ci now
