+++ tests/assets/airlock/mongodbcluster.yaml
@@ -1,0 +1,19 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: airlock-connection-string
+  namespace: airlock-system
+type: Opaque
+stringData:
+  connectionString: "mongodb://root:root@mongo.mongo.svc.cluster.local:27017/?directConnection=true"
+---
+apiVersion: airlock.cloud.rocket.chat/v1alpha1
+kind: MongoDBCluster
+metadata:
+  name: airlock-test
+spec:
+  connectionSecret: airlock-connection-string
+  hostTemplate: "mongo.mongo.svc.cluster.local"
+  optionsTemplate: "?directConnection=true"
+  prefixTemplate: mongodb
+---
