+++ tests/assets/mongo/mongo-pod.yaml
@@ -1,0 +1,18 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: mongo
+  namespace: mongo
+  labels:
+    app: mongo
+spec:
+  containers:
+    - name: mongo
+      image: mongo:7.0
+      env:
+        - name: MONGO_INITDB_ROOT_USERNAME
+          value: "root"
+        - name: MONGO_INITDB_ROOT_PASSWORD
+          value: "root"
+      ports:
+        - containerPort: 27017
